edit view dan bug

+++ resources/views/petugas.blade.php
@@ -26,11 +26,11 @@
                     <table id="petugas" class="table table-striped table-bordered" style="width:100%">
                         <thead>
                             <tr>
-                                <th>No</th>
-                                <th>Username</th>
-                                <th>Role</th>
+                                <th class="text-center">No</th>
+                                <th class="text-center">Username</th>
+                                <th class="text-center">Role</th>
                                 @if(Auth::user()->role == 'admin')
-                                    <th>Action</th>
+                                    <th class="text-center">Action</th>
                                 @endif
                             </tr>
                         </thead>
@@ -38,22 +38,26 @@
                             @php $n=1 @endphp
                             @foreach($petugas as $data)
                             <tr>
-                                <td>{{$n++}}</td>
-                                <td>{{ $data->username }}</td>
-                                <td>
-                                    @if($data->role == 'admin')
-                                    <span class="label label-danger">{{ $data->role }}</span>
-                                    @else
-                                    <span class="label label-primary">{{ $data->role }}</span>
-                                    @endif
-                                </td>
-                                @if(Auth::user()->role == 'admin')
-                                    <td>
-                                        <button type="button" class="btn btn-warning" data-toggle="modal" data-target="#editPetugas" onclick="getData('{{ $data->id }}')">
-                                          Edit
-                                        </button>
-                                        <button class="btn btn-danger" onclick="deleteData('{{ $data->id }}')">Delete</button>
+                                @if($data->username != 'system')
+                                    <td class="text-center">{{$n++}}</td>
+                                    <td>{{ $data->username }}</td>
+                                    <td class="text-center">
+                                        @if($data->role == 'admin')
+                                        <span class="label label-danger">{{ $data->role }}</span>
+                                        @else
+                                        <span class="label label-primary">{{ $data->role }}</span>
+                                        @endif
                                     </td>
+                                    @if(Auth::user()->role == 'admin')
+                                        <td class="text-center">
+                                            <button type="button" class="btn btn-primary" data-toggle="modal" data-target="#editPetugas" onclick="getData('{{ $data->id }}')">Change Username</button>
+                                            <button class="btn btn-info" onclick="changeRole('{{ $data->id }}','{{ $data->role }}')">Change Role</button>
+                                            <button class="btn btn-warning" onclick="resetPassword('{{ $data->id }}')">
+                                              Reset Password
+                                            </button>
+                                            <button class="btn btn-danger" onclick="deleteData('{{ $data->id }}')">Delete</button>
+                                        </td>
+                                    @endif
                                 @endif
                             </tr>
                             @endforeach
@@ -81,10 +85,6 @@
                     <input type="text" class="form-control" id="username" placeholder="Username">
                 </div>
                 <div class="form-group">
-                    <label for="password" class="col-form-label">Password</label>
-                    <input type="password" class="form-control" id="password" placeholder="Password">
-                </div>
-                <div class="form-group">
                     <label for="role">Role</label>
                     <select id="role" class="form-control">
                         <option value="petugas">Petugas</option>
@@ -101,33 +101,22 @@
       </div>
     </div>
 
-    <!-- EDIT MODAL -->
+    <!-- ADD MODAL -->
     <div class="modal fade" id="editPetugas" tabindex="-1" role="dialog" aria-labelledby="editPetugasCenterTitle" aria-hidden="true">
       <div class="modal-dialog modal-dialog-centered" role="document">
         <div class="modal-content">
           <div class="modal-header">
-            <h5 class="modal-title" id="editPetugasCenterTitle">Form Edit Data Petugas</h5>
+            <h5 class="modal-title" id="EditPetugasCenterTitle">Form Edit Username Petugas</h5>
             <button type="button" class="close" data-dismiss="modal" aria-label="Close">
               <span aria-hidden="true">&times;</span>
             </button>
           </div>
           <div class="modal-body">
-            <form id="form_edit">
-                <input type="hidden" class="form-control" id="editID">
+            <form id="form_add">
                 <div class="form-group">
+                    <input type="hidden" class="form-control" id="editID">
                     <label for="editUsername" class="col-form-label">Username</label>
                     <input type="text" class="form-control" id="editUsername" placeholder="Username">
-                </div>
-                <div class="form-group">
-                    <label for="editPassword" class="col-form-label">Password</label>
-                    <input type="password" class="form-control" id="editPassword" placeholder="Password">
-                </div>
-                <div class="form-group">
-                    <label for="editRole">Role</label>
-                    <select id="editRole" class="form-control">
-                        <option value="petugas">Petugas</option>
-                        <option value="admin">Admin</option>
-                    </select>
                 </div>
             </form>
           </div>
@@ -138,6 +127,7 @@
         </div>
       </div>
     </div>
+
 </div>
 @endsection
 
@@ -147,41 +137,59 @@
         $('#petugas').DataTable();
     });
 
+    // GET DATA PETUGAS
     function getData(id) {
         $.get('{{ route("petugas.index") }}/'+id, function(data) {
             $('#editUsername').val(data.username);
-            $('#editRole').val(data.role);
             $('#editID').val(data.id); 
         });
     }
 
-    $('#saveAdd').click(function(e) {
-        e.preventDefault();
-        if($('#username').val() == '' || $('#password').val() == '') {
-            alert("Username atau Password tidak boleh kosong!");
-        } else {
-            $.ajax({  
-                url: '{{ route("petugas.index") }}',  
-                type: 'POST',  
-                dataType: 'json',  
-                data: {
-                    username: $('#username').val(),
-                    password: $('#password').val(),
-                    role: $('#role').val()
-                },  
-                success: function (data) {
-                    if(data.status == 0) {
-                        alert(data.msg);
-                    } else {
-                        alert('Data berhasil ditambah.');
-                        refreshForm();
-                        $("#petugas").load(window.location + " #petugas");
-                    }
-                }
-            });
+    $('#addPetugas').bind('keydown', function(e) {
+        if (e.keyCode == 13) {
+            e.preventDefault();
         }
     });
 
+    $('#editPetugas').bind('keydown', function(e) {
+        if (e.keyCode == 13) {
+            e.preventDefault();
+        }
+    });
+
+    $('#saveAdd').click(function(e) {
+        e.preventDefault();
+        if($('#username').val() == '') {
+            alert("Username tidak boleh kosong!");
+        } else {
+            if($('#username').val().length < 5){
+                alert('Username minimal 5 karakter!');
+            }else if($('#username').val().includes(' ') == true){
+                alert('Username tidak boleh menggunakan sepasi!');
+            }else{
+                $.ajax({  
+                    url: '{{ route("petugas.index") }}',  
+                    type: 'POST',  
+                    dataType: 'json',  
+                    data: {
+                        username: $('#username').val(),
+                        role: $('#role').val()
+                    },  
+                    success: function (data) {
+                        if(data.status == 0) {
+                            alert(data.msg);
+                        } else {
+                            alert('Data berhasil ditambah.');
+                            refreshForm();
+                            $("#petugas").load(window.location + " #petugas");
+                        }
+                    }
+                });
+            }
+        }
+    });
+
+    // EDIT DATA PETUGAS
     $('#saveEdit').click(function(e) {
         e.preventDefault();
 
@@ -190,9 +198,7 @@
             type: 'PATCH',  
             dataType: 'json',  
             data: {
-                username: $('#editUsername').val(),
-                password: $('#editPassword').val(),
-                role: $('#editRole').val()
+                username: $('#editUsername').val()
             },  
             success: function (data) {
                 if(data.status == 0){
@@ -205,8 +211,52 @@
         });
     });
 
+    function resetPassword(id){
+        $.ajax({  
+            url: '{{ route("petugas.index") }}/'+id,  
+            type: 'PATCH',  
+            dataType: 'json',  
+            data: {
+                reset: true
+            },  
+            success: function (data) {
+                if(data.status == 0){
+                    alert(data.msg);
+                }else{
+                    alert('Password berhasil direset.');
+                    $("#petugas").load(window.location + " #petugas");
+                }
+            }
+        });
+    }
+
+    function changeRole(id, roleName){
+        $.ajax({  
+            url: '{{ route("petugas.index") }}/'+id,  
+            type: 'PATCH',  
+            dataType: 'json',  
+            data: {
+                role: roleName
+            },  
+            success: function (data) {
+                if(data.status == 0){
+                    alert(data.msg);
+                }else{
+                    alert('Role berhasil diganti.');
+                    window.location.reload();
+                    let userRole = '{{Auth::user()->role}}';
+                    if( userRole == 'petugas'){
+                        window.location.href='{{ route("home") }}';
+                    }else{
+                        $("#petugas").load(window.location + " #petugas");
+                    }
+                }
+            }
+        });
+    }
+
     function deleteData(id) {
-        let conf = confirm("Apakah anda yakin data ini akan dihapus ?");
+        let conf = confirm("Jika data Petugas dihapus, maka data di Absen yang terhubung akan terhapus. Apakah Anda yakin data ini akan dihapus ?");
         if(conf){
             $.ajax({  
                 url: '{{ route("petugas.index") }}/'+id,  
@@ -226,7 +276,6 @@
 
     function refreshForm() {
         $('#username').val('');
-        $('#password').val('');
         $('#role').val('petugas');
     }
 </script>
